2.1.4

+++ app/models/IconPack.ts
@@ -1,6 +1,6 @@
 import axios from 'axios';
 import electron from 'electron';
-import tmp from 'tmp';
+import rimraf from 'rimraf';
 import unzipper from 'unzipper';
 import fs from 'fs-extra';
 import path from 'path';
@@ -8,15 +8,24 @@ import log from 'electron-log';
 import settingsUtil from '../settings/Settings';
 import logger from 'electron-log';
 import { PackMeta } from './PackMeta';
+import { promisify } from 'util';
 
-const {ipcRenderer} = electron;
+const { ipcRenderer } = electron;
 
 axios.defaults.adapter = require('axios/lib/adapters/http');
 
 export default abstract class IconPack {
   meta: PackMeta;
+  static tempDir = path.resolve((electron.app || electron.remote.app).getPath('userData'), 'temp');
   constructor(meta: PackMeta) {
     this.meta = meta;
+  }
+
+  static async cleanTemp() {
+    const rm = promisify(rimraf);
+    logger.info(`Cleaning up temp directory ${IconPack.tempDir}`);
+    await rm(IconPack.tempDir);
+    await fs.promises.mkdir(IconPack.tempDir);
   }
 
   /**
@@ -56,7 +65,7 @@ export default abstract class IconPack {
    */
   private async extractZip(zipPath: string) {
     return new Promise((resolve, reject) => {
-      const tmpDir =  {name: path.resolve((electron.app || electron.remote.app).getPath('userData'),`${this.meta.id}`)};
+      const tmpDir = { name: path.resolve(IconPack.tempDir,`${Date.now()}_${this.meta.id}`) };
       fs.createReadStream(zipPath)
         .pipe(unzipper.Extract({ path: tmpDir.name }))
         .on('close', () => {
@@ -75,7 +84,7 @@ export default abstract class IconPack {
    */
   private async downloadZip(onProgress?: Function): Promise<Buffer> {
     const url = await this.getZipUrl();
-    const zip =  {name: path.resolve((electron.app || electron.remote.app).getPath('userData'),`${this.meta.id}.zip`)};
+    const zip = { name: path.resolve(IconPack.tempDir, `${Date.now()}_${this.meta.id}.zip`) };
     return new Promise((resolve, reject) => {
       ipcRenderer.send('downloadFile', {
         outputLocation: zip.name,
@@ -103,7 +112,7 @@ export default abstract class IconPack {
     log.debug('Extracting Zip');
     const extractDir = await this.extractZip(zipPath.name);
     log.debug(`Extracted to ${extractDir.name}`);
-    return {zipPath, extractPath: extractDir};
+    return { zipPath, extractPath: extractDir };
   }
 
   /**
@@ -128,11 +137,27 @@ export default abstract class IconPack {
       'Icons'
     );
 
-    const paths = await this.downloadAndExtract(onProgress);
-    await this.copyFilesTo(`${paths.extractPath.name}/Pack`, dbdIconsPath, opts);
-    log.debug('Files copied!');
-    await this.saveInstalledPackId();
-    await fs.remove(paths.extractPath.name);
-    await fs.remove(paths.zipPath.name);
+    let paths: any;
+
+    try {
+      paths = await this.downloadAndExtract(onProgress);
+      await this.copyFilesTo(`${paths.extractPath.name}/Pack`, dbdIconsPath, opts);
+      await this.saveInstalledPackId();
+    } finally {
+      if (paths?.extractPath) {
+        try {
+          logger.info(`Removing ${paths.extractPath.name}`);
+          await fs.remove(paths.extractPath.name);
+        } catch (e) { }
+      }
+      if (paths?.zipPath) {
+        try {
+          logger.info(`Removing ${paths.zipPath.name}`);
+          await fs.remove(paths.zipPath.name);
+        } catch (e) { }
+      }
+    }
+
+
   }
 }
